More ANSI changes
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1541 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/nethack3.0.d/include/decl.h
@@ -8,7 +8,7 @@
 #define E extern
 
 E int bases[];
-E int warnlevel;	/* defined in mon.c */
+E uchar warnlevel;	/* defined in mon.c */
 E int occtime;
 E int nroom;
 E int multi;
